cleaning, api views and templateviews splitup
